+++ src/test/java/com/crud/theatre/Facade/SpectacleFacadeTest.java
@@ -96,7 +96,7 @@ public class SpectacleFacadeTest {
 
         //then
         actors.stream().forEach(actorDto -> {
-            assertEquals(1l, actorDto.getId().longValue());
+            assertEquals(1l, actorDto.getId());
             assertEquals("firstName test", actorDto.getFirstName());
             assertEquals("lastName test", actorDto.getLastName());
         });
